Reorganisation into scripts folder and split examples and solutions
