+++ scripts/example-tests.log
@@ -1,0 +1,4 @@
+---Executing '3dplot.py'---
+---END '3dplot.py'---
+---Executing '3dvec.py'---
+---END '3dvec.py'---
